+++ example/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_osm_plugin_example/src/search_example.dart';
 
 //import 'src/adv_home/home_example.dart';
@@ -7,7 +8,8 @@ import 'src/home/home_example.dart';
 import 'src/search_example.dart';
 import 'web_test_osm.dart';
 
-void main() {
+void main() async {
+  await dotenv.load(fileName: ".env");
   WidgetsFlutterBinding.ensureInitialized();
   if (kIsWeb) {
     runApp(WebApp());
@@ -21,7 +23,6 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       initialRoute: "/home",
-
       routes: {
         "/home": (ctx) => MainExample(),
         //"/adv-home": (ctx) => AdvandedMainExample(),