+++ flutter_osm_interface/lib/src/osm_controller/osm_controller.dart
@@ -33,6 +33,12 @@ abstract class IBaseOSMController {
     double? angle,
   });
 
+  Future<void> changeMarker({
+    required GeoPoint oldLocation,
+    required GeoPoint newLocation,
+    MarkerIcon? newMarkerIcon,
+  });
+
   /// removeMarker
   /// remove marker from map of position
   ///
@@ -248,6 +254,7 @@ abstract class IBaseOSMController {
   Future<void> limitArea(
     BoundingBox box,
   );
+
   /// removeLimitArea
   ///
   /// this method will remove the region limitation for camera movement