+++ flutter_osm_interface/lib/src/osm_interface.dart
@@ -107,6 +107,13 @@ abstract class MobileOSMPlatform extends OSMPlatform {
     GlobalKey? globalKeyIcon,
   });
 
+  Future<void> changeMarker(
+    int idOSM,
+    GeoPoint oldLocation,
+    GeoPoint newLocation, {
+    GlobalKey? globalKeyIcon,
+  });
+
   Future<void> setIconMarker(
     int idOSM,
     GeoPoint point,